+++ src/__test__/unit/AbilityRuleSet.test.ts
@@ -1,7 +1,7 @@
-import AbilityRule from '../../AbilityRule';
-import AbilityRuleSet from '../../AbilityRuleSet';
-import AbilityMatch from '../../AbilityMatch';
-import AbilityCompare from '../../AbilityCompare';
+import AbilityRule from '../../core/AbilityRule';
+import AbilityRuleSet from '../../core/AbilityRuleSet';
+import AbilityMatch from '../../core/AbilityMatch';
+import AbilityCompare from '../../core/AbilityCompare';
 
 describe('AbilityRuleSet', () => {
   describe('constructor', () => {